+++ list/ft_lstdelone.c
@@ -21,6 +21,5 @@ void	ft_lstdelone(t_list *lst, void (*del)(void *))
 	{	
 		del(lst->content);
 		free(lst);
-		lst = lst->next;
 	}
 }